+++ firmware/BlockBoxController/ModuleInterface/Src/module_interface.cpp
@@ -8,7 +8,7 @@
 #include "module_interface.h"
 
 
-uint8_t ModuleInterface::ReadRegister8(uint8_t reg_addr) {
+uint8_t ModuleInterface::ReadRegister8(uint16_t reg_addr) {
   uint8_t value;
 
   this->ReadRegister(reg_addr, &value, 1);
@@ -16,7 +16,7 @@ uint8_t ModuleInterface::ReadRegister8(uint8_t reg_addr) {
   return value;
 }
 
-uint16_t ModuleInterface::ReadRegister16(uint8_t reg_addr) {
+uint16_t ModuleInterface::ReadRegister16(uint16_t reg_addr) {
   uint16_t value;
 
   this->ReadRegister(reg_addr, (uint8_t*)&value, 2);
@@ -24,7 +24,7 @@ uint16_t ModuleInterface::ReadRegister16(uint8_t reg_addr) {
   return value;
 }
 
-uint32_t ModuleInterface::ReadRegister32(uint8_t reg_addr) {
+uint32_t ModuleInterface::ReadRegister32(uint16_t reg_addr) {
   uint32_t value;
 
   this->ReadRegister(reg_addr, (uint8_t*)&value, 4);
@@ -35,7 +35,7 @@ uint32_t ModuleInterface::ReadRegister32(uint8_t reg_addr) {
 
 //helper function for queueing variable length reads or writes
 static inline void _ModuleInterface_QueueVarTransfer(std::deque<ModuleTransferQueueItem*>& queue, ModuleTransferQueueItem* new_transfer, ModuleTransferType type,
-                                                     uint8_t reg_addr, uint8_t* buf, uint16_t length, ModuleTransferCallback&& callback) {
+                                                     uint16_t reg_addr, uint8_t* buf, uint16_t length, ModuleTransferCallback&& callback) {
   if (buf == NULL || length == 0) {
     delete new_transfer;
     throw std::invalid_argument("ModuleInterface transfers require non-null buffer and nonzero length");
@@ -60,7 +60,7 @@ static inline void _ModuleInterface_QueueVarTransfer(std::deque<ModuleTransferQu
 
 //helper function for queueing 8/16/32-bit reads or writes
 static inline void _ModuleInterface_QueueShortTransfer(std::deque<ModuleTransferQueueItem*>& queue, ModuleTransferQueueItem* new_transfer, ModuleTransferType type,
-                                                       uint8_t reg_addr, uint32_t value, uint16_t length, ModuleTransferCallback&& callback) {
+                                                       uint16_t reg_addr, uint32_t value, uint16_t length, ModuleTransferCallback&& callback) {
   //configure transfer for the short read (reading into the internal value buffer) or write (copy value into the internal value buffer, write from there)
   new_transfer->type = type;
   new_transfer->reg_addr = reg_addr;
@@ -78,56 +78,56 @@ static inline void _ModuleInterface_QueueShortTransfer(std::deque<ModuleTransfer
   }
 }
 
-void ModuleInterface::ReadRegisterAsync(uint8_t reg_addr, uint8_t* buf, uint16_t length, ModuleTransferCallback&& callback) {
+void ModuleInterface::ReadRegisterAsync(uint16_t reg_addr, uint8_t* buf, uint16_t length, ModuleTransferCallback&& callback) {
   _ModuleInterface_QueueVarTransfer(this->queued_transfers, this->CreateTransferQueueItem(), TF_READ, reg_addr, buf, length, std::move(callback));
   this->StartQueuedAsyncTransfer();
 }
 
-void ModuleInterface::ReadRegister8Async(uint8_t reg_addr, ModuleTransferCallback&& callback) {
+void ModuleInterface::ReadRegister8Async(uint16_t reg_addr, ModuleTransferCallback&& callback) {
   _ModuleInterface_QueueShortTransfer(this->queued_transfers, this->CreateTransferQueueItem(), TF_READ, reg_addr, 0, 1, std::move(callback));
   this->StartQueuedAsyncTransfer();
 }
 
-void ModuleInterface::ReadRegister16Async(uint8_t reg_addr, ModuleTransferCallback&& callback) {
+void ModuleInterface::ReadRegister16Async(uint16_t reg_addr, ModuleTransferCallback&& callback) {
   _ModuleInterface_QueueShortTransfer(this->queued_transfers, this->CreateTransferQueueItem(), TF_READ, reg_addr, 0, 2, std::move(callback));
   this->StartQueuedAsyncTransfer();
 }
 
-void ModuleInterface::ReadRegister32Async(uint8_t reg_addr, ModuleTransferCallback&& callback) {
+void ModuleInterface::ReadRegister32Async(uint16_t reg_addr, ModuleTransferCallback&& callback) {
   _ModuleInterface_QueueShortTransfer(this->queued_transfers, this->CreateTransferQueueItem(), TF_READ, reg_addr, 0, 4, std::move(callback));
   this->StartQueuedAsyncTransfer();
 }
 
 
-void ModuleInterface::WriteRegister8(uint8_t reg_addr, uint8_t value) {
+void ModuleInterface::WriteRegister8(uint16_t reg_addr, uint8_t value) {
   this->WriteRegister(reg_addr, &value, 1);
 }
 
-void ModuleInterface::WriteRegister16(uint8_t reg_addr, uint16_t value) {
+void ModuleInterface::WriteRegister16(uint16_t reg_addr, uint16_t value) {
   this->WriteRegister(reg_addr, (uint8_t*)&value, 2);
 }
 
-void ModuleInterface::WriteRegister32(uint8_t reg_addr, uint32_t value) {
+void ModuleInterface::WriteRegister32(uint16_t reg_addr, uint32_t value) {
   this->WriteRegister(reg_addr, (uint8_t*)&value, 4);
 }
 
 
-void ModuleInterface::WriteRegisterAsync(uint8_t reg_addr, const uint8_t* buf, uint16_t length, ModuleTransferCallback&& callback) {
+void ModuleInterface::WriteRegisterAsync(uint16_t reg_addr, const uint8_t* buf, uint16_t length, ModuleTransferCallback&& callback) {
   _ModuleInterface_QueueVarTransfer(this->queued_transfers, this->CreateTransferQueueItem(), TF_WRITE, reg_addr, (uint8_t*)buf, length, std::move(callback));
   this->StartQueuedAsyncTransfer();
 }
 
-void ModuleInterface::WriteRegister8Async(uint8_t reg_addr, uint8_t value, ModuleTransferCallback&& callback) {
+void ModuleInterface::WriteRegister8Async(uint16_t reg_addr, uint8_t value, ModuleTransferCallback&& callback) {
   _ModuleInterface_QueueShortTransfer(this->queued_transfers, this->CreateTransferQueueItem(), TF_WRITE, reg_addr, (uint32_t)value, 1, std::move(callback));
   this->StartQueuedAsyncTransfer();
 }
 
-void ModuleInterface::WriteRegister16Async(uint8_t reg_addr, uint16_t value, ModuleTransferCallback&& callback) {
+void ModuleInterface::WriteRegister16Async(uint16_t reg_addr, uint16_t value, ModuleTransferCallback&& callback) {
   _ModuleInterface_QueueShortTransfer(this->queued_transfers, this->CreateTransferQueueItem(), TF_WRITE, reg_addr, (uint32_t)value, 2, std::move(callback));
   this->StartQueuedAsyncTransfer();
 }
 
-void ModuleInterface::WriteRegister32Async(uint8_t reg_addr, uint32_t value, ModuleTransferCallback&& callback) {
+void ModuleInterface::WriteRegister32Async(uint16_t reg_addr, uint32_t value, ModuleTransferCallback&& callback) {
   _ModuleInterface_QueueShortTransfer(this->queued_transfers, this->CreateTransferQueueItem(), TF_WRITE, reg_addr, value, 4, std::move(callback));
   this->StartQueuedAsyncTransfer();
 }
